+++ src/components/Main/MainView/QallView/QallView.vue
